LinearSlide and Intake subsystem updated

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/drive/subsystems/IntakeSubsystem.java
@@ -17,7 +17,7 @@ public class IntakeSubsystem {
         Intake.setPower(0);
     }
 
-    public void setPower(double power) {
+    public void setPower(int power) {
         Intake.setPower(power);
     }
